requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
python-dotenv>=1.0
dj-database-url>=1.0
gunicorn>=20.1
whitenoise>=6.5
psycopg2-binary>=2.9
qrcode>=7.3
Pillow>=9.0
Django>=4.2
djangorestframework
Pillow
qrcode
psycopg2-binary
python-dotenv
dj-database-url
gunicorn